# packages/sdk/openapi/trinacria-cms.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trinacria CMS Playground API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "operationId": "listApiKeys",
        "summary": "List issued API keys",
        "tags": ["Api Keys"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["publishable", "secret", "service"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "revoked"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "id",
                          "keyPrefix",
                          "secretPreview",
                          "name",
                          "kind",
                          "status",
                          "roleCodes",
                          "permissionKeys",
                          "policyRules",
                          "createdAt",
                          "updatedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "keyPrefix": {
                            "type": "string"
                          },
                          "secretPreview": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": ["publishable", "secret", "service"]
                          },
                          "status": {
                            "type": "string",
                            "enum": ["active", "revoked"]
                          },
                          "roleCodes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "permissionKeys": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "policyRules": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "additionalProperties": false,
                              "required": ["effect", "permissionPattern", "conditions"],
                              "properties": {
                                "effect": {
                                  "type": "string",
                                  "enum": ["allow", "deny"]
                                },
                                "permissionPattern": {
                                  "type": "string"
                                },
                                "conditions": {
                                  "type": "array",
                                  "items": {
                                    "type": "string",
                                    "enum": ["resource_id_required", "resource_id_equals_subject"]
                                  }
                                }
                              }
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "lastUsedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "revokedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pluginId": {
                          "enum": ["core-pack"]
                        },
                        "count": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": ["data"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createApiKey",
        "summary": "Issue a new API key",
        "tags": ["Api Keys"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "kind": {
                    "type": "string",
                    "enum": ["publishable", "secret", "service"]
                  },
                  "roleCodes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "permissionKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "policyRules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": ["effect", "permissionPattern", "conditions"],
                      "properties": {
                        "effect": {
                          "type": "string",
                          "enum": ["allow", "deny"]
                        },
                        "permissionPattern": {
                          "type": "string"
                        },
                        "conditions": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": ["resource_id_required", "resource_id_equals_subject"]
                          }
                        }
                      }
                    }
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Issued API key with one-time secret",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": ["record", "apiKey"],
                      "properties": {
                        "record": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "id",
                            "keyPrefix",
                            "secretPreview",
                            "name",
                            "kind",
                            "status",
                            "roleCodes",
                            "permissionKeys",
                            "policyRules",
                            "createdAt",
                            "updatedAt"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "keyPrefix": {
                              "type": "string"
                            },
                            "secretPreview": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "kind": {
                              "type": "string",
                              "enum": ["publishable", "secret", "service"]
                            },
                            "status": {
                              "type": "string",
                              "enum": ["active", "revoked"]
                            },
                            "roleCodes": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "permissionKeys": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "policyRules": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "additionalProperties": false,
                                "required": ["effect", "permissionPattern", "conditions"],
                                "properties": {
                                  "effect": {
                                    "type": "string",
                                    "enum": ["allow", "deny"]
                                  },
                                  "permissionPattern": {
                                    "type": "string"
                                  },
                                  "conditions": {
                                    "type": "array",
                                    "items": {
                                      "type": "string",
                                      "enum": ["resource_id_required", "resource_id_equals_subject"]
                                    }
                                  }
                                }
                              }
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "lastUsedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "expiresAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "revokedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        },
                        "apiKey": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pluginId": {
                          "enum": ["core-pack"]
                        },
                        "count": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": ["data"],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "API key configuration invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pluginId": {
                          "enum": ["core-pack"]
                        },
                        "count": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "get": {
        "operationId": "getApiKeyById",
        "summary": "Get one API key metadata record",
        "tags": ["Api Keys"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "id",
                        "keyPrefix",
                        "secretPreview",
                        "name",
                        "kind",
                        "status",
                        "roleCodes",
                        "permissionKeys",
                        "policyRules",
                        "createdAt",
                        "updatedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "keyPrefix": {
                          "type": "string"
                        },
                        "secretPreview": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string",
                          "enum": ["publishable", "secret", "service"]
                        },
                        "status": {
                          "type": "string",
                          "enum": ["active", "revoked"]
                        },
                        "roleCodes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "permissionKeys": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "policyRules": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": ["effect", "permissionPattern", "conditions"],
                            "properties": {
                              "effect": {
                                "type": "string",
                                "enum": ["allow", "deny"]
                              },
                              "permissionPattern": {
                                "type": "string"
                              },
                              "conditions": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "enum": ["resource_id_required", "resource_id_equals_subject"]
                                }
                              }
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "lastUsedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "revokedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pluginId": {
                          "enum": ["core-pack"]
                        },
                        "count": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": ["data"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pluginId": {
                          "enum": ["core-pack"]
                        },
                        "count": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste access token as `Bearer <token>`"
      },
      "pluginCallerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cms-plugin-signature",
        "description": "Signed plugin caller flow. Requests also require x-cms-plugin-id, x-cms-plugin-ts, and x-cms-plugin-nonce headers."
      }
    }
  }
}